feat: TimeService create/delete 시 검증 로직 추가

+++ src/main/java/roomescape/service/TimeService.java
@@ -5,6 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.Time;
 import roomescape.dto.TimeRequest;
 import roomescape.dto.TimeResponse;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.TimeRepository;
 
 import java.util.List;
@@ -12,9 +13,11 @@ import java.util.stream.Collectors;
 
 @Service
 public class TimeService {
+    private final ReservationRepository reservationRepository;
     private final TimeRepository timeRepository;
 
-    public TimeService(TimeRepository timeRepository) {
+    public TimeService(ReservationRepository reservationRepository, TimeRepository timeRepository) {
+        this.reservationRepository = reservationRepository;
         this.timeRepository = timeRepository;
     }
 
@@ -26,13 +29,20 @@ public class TimeService {
 
     @Transactional
     public TimeResponse createTime(TimeRequest request) {
-        Time time = new Time(null, request.time());
+        timeRepository.findByTime(request.time())
+                .ifPresent(t -> { throw new IllegalStateException("이미 존재하는 시간입니다."); });
+
+        Time time = Time.create(request.time());
         Time savedTime = timeRepository.save(time);
         return TimeResponse.from(savedTime);
     }
 
     @Transactional
     public void deleteTime(Long id) {
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new IllegalStateException("예약이 존재하는 시간은 삭제할 수 없습니다.");
+        }
+
         timeRepository.deleteById(id);
     }
 }
